Change the license from LGPL-2.1-or-later to Apache-2.0
The initial version of subuidless adopted LGPL-2.1-or-later because
it was expected to be going to be a shared object library that is
called by GPL2-licensed crun.

However, the current version of subuidless is implemented as a socket
receiver process rather than a shared object library, and probably continue
to being a socket receiver process so that it can be executed in the parent
namespace.

(Side topic: we need an infra to allow using multiple socket receivers)

And yet, even if it was a shared object library, probably it does not need to be LGPL:
https://github.com/containers/crun/pull/438

So this commit changes the license to Apache-2.0 for consistency with
other container software in the ecosystem.

Signed-off-by: Akihiro Suda <akihiro.suda.cz@hco.ntt.co.jp>